refactor(error): 更新异常类：构造器优化、错误码调整 fix(i18n):  使用应用默认编码读取配置文件，避免乱码 feat(log): 添加调用日志拦截器、添加记录 throw 语法糖方法 feat(util): 添加异常格式化工具、添加断言工具、ContextUtil 标注返回值是否可空 fix(test): 修复单测 feat(validator): 添加校验工具类

+++ shoulder-build/shoulder-base/shoulder-core/src/main/java/org/shoulder/core/dto/response/ListResult.java
@@ -22,17 +22,17 @@ public class ListResult<T> implements Serializable {
     private static final long serialVersionUID = -3134782461635924904L;
     //@Schema(name = "数据总数")
     @ApiModelProperty(value = "数据总数", dataType = "long", required = true, example = "4", position = 0)
-    private int total = 0;
+    private long total = 0L;
 
     //@Schema(name = "列表数据")
     @ApiModelProperty(value = "列表数据", dataType = "List", required = true, example = "[1,2,3,4]", position = 1)
     private List<T> list;
 
-    public int getTotal() {
+    public long getTotal() {
         return total;
     }
 
-    public void setTotal(int total) {
+    public void setTotal(long total) {
         this.total = total;
     }
